Move find_in_file to utils module
